created Header component
