+++ src/components/Header.js
@@ -1,0 +1,29 @@
+import { Box, Container, Grid, Typography } from '@mui/material';
+
+function Header() {
+  const containerStyle = {
+    backgroundImage:
+      'linear-gradient(53deg, rgba(78,43,233,1) 20%, rgba(121,9,111,1) 100%)',
+    width: '100%',
+    position: 'absolute',
+    height: '200px',
+    zIndex: '-1',
+  };
+
+  return (
+    <Grid item xs={12} md={8} lg={8}>
+      <Container sx={containerStyle}></Container>
+      <Box sx={{ position: 'relative' }}>
+        <Typography
+          variant="h3"
+          textAlign="center"
+          paddingTop="30px"
+          color="white">
+          My Tasks
+        </Typography>
+      </Box>
+    </Grid>
+  );
+}
+
+export default Header;
