+++ src/components/Tasks.js
@@ -1,34 +1,16 @@
-import { Box, Container, Grid, Typography } from '@mui/material';
+import { Grid } from '@mui/material';
 import { useSelector } from 'react-redux';
 import TaskForm from './TaskForm';
-function Tasks() {
-  const containerStyle = {
-    backgroundImage:
-      'linear-gradient(53deg, rgba(78,43,233,1) 20%, rgba(121,9,111,1) 100%)',
-    width: '100%',
-    position: 'absolute',
-    height: '200px',
-    zIndex: '-1',
-  };
+import Header from './Header';
 
+function Tasks() {
   const tasks = useSelector((state) => state.tasks);
 
   return (
     <>
       <Grid alignItems="center" justifyContent="center" container spacing={0}>
-        {/* header */}
-        <Grid item xs={12} md={8} lg={8}>
-          <Box sx={{ position: 'relative' }}>
-            <Container sx={containerStyle}></Container>
-            <Typography
-              variant="h3"
-              textAlign="center"
-              paddingTop="30px"
-              color="white">
-              My Tasks
-            </Typography>
-          </Box>
-        </Grid>
+        <Header />
+
         {/* task form */}
         <Grid item xs={12} md={8} lg={8}>
           <TaskForm />
